BUG: Fix migrations not working on Postgres
This requires me to redo EVERY migration. All 52 of them. Either that,
or remove them, which is what I did.

Hopefully this teaches me that 4 billion characters of precision really
aren't needed. I can't really see someone having 10^1000 OK, which is
the new limit.

Sorry if this caused any inconveniences, but it was required. Shit
happens unfortunatelly, and sometimes that's due to your incompetence.